+++ disp.h
@@ -34,7 +34,7 @@ class Display {
       int16_t y=(320-256), int16_t h=256, int16_t xlab=0, int16_t noise=0);  
     void ShowChartPlus(const double *pdVals, int16_t nvals, 
       int16_t y=(320-256), int16_t h=256, int16_t xlab=0, int16_t noise=0);
-    void ShowChartPlusMax(const double *pdVals, int16_t nvals, 
+    void ShowChartPlusMax(const int16_t *pdVals, int16_t nvals, 
       int16_t y=(320-256), int16_t h=256, int16_t xlab=0, int16_t vmax=100, int16_t noise=0);  
   protected:
     void BufLen();
